refactor: remove sortingBox from filters, create folder for it
